feat(vhd-lib/parseVhdStream): new public method
Extracted from `createVhdDirectoryFromStream`
