+++ packages/vhd-lib/src/index.js
@@ -8,6 +8,7 @@ export { createVhdDirectoryFromStream } from './createVhdDirectoryFromStream'
 export { default as mergeVhd } from './merge'
 export { default as peekFooterFromVhdStream } from './peekFooterFromVhdStream'
 export { openVhd } from './openVhd'
+export { parseVhdStream } from './parseVhdStream'
 export { VhdAbstract } from './Vhd/VhdAbstract'
 export { VhdDirectory } from './Vhd/VhdDirectory'
 export { VhdFile } from './Vhd/VhdFile'
